renamed the handler and related files to huggingface_api_handler
