+++ mindsdb/integrations/handlers/huggingface_api_handler/__init__.py
@@ -2,14 +2,14 @@ from mindsdb.integrations.libs.const import HANDLER_TYPE
 
 from mindsdb.integrations.handlers.autosklearn_handler.__about__ import __version__ as version, __description__ as description
 try:
-    from .huggingface_inference_api_handler import HuggingFaceInferenceAPIHandler as Handler
+    from .huggingface_api_handler import HuggingFaceInferenceAPIHandler as Handler
     import_error = None
 except Exception as e:
     Handler = None
     import_error = e
 
-title = 'Hugging Face Inference API'
-name = 'huggingface_inference_api'
+title = 'Hugging Face API'
+name = 'huggingface_api'
 type = HANDLER_TYPE.ML
 permanent = True
